backend/Auth/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
black==23.1.0
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
Deprecated==1.2.13
Django==4.1.7
django-cors-headers==3.13.0
django-csp==3.7
django-environ==0.9.0
django-oauth-toolkit==2.2.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
idna==3.4
jwcrypto==1.4.2
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
psycopg==3.1.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==1.7.1
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
requirements-parser==0.5.0
sentry-sdk==1.15.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
types-docutils==0.19.1.6
types-setuptools==67.4.0.1
typing_extensions==4.5.0
urllib3==1.26.14
webencodings==0.5.1
wrapt==1.14.1
